refactor: removed multi faceted abstraction smell by splitting tasks of DirectoryScanner class into FileProcessor, FileTypeDetector and DirectorScanner class
