+++ impexp-core/src/main/java/org/citydb/core/file/input/FileProcessor.java
@@ -1,0 +1,131 @@
+/*
+ * 3D City Database - The Open Source CityGML Database
+ * https://www.3dcitydb.org/
+ *
+ * Copyright 2013 - 2024
+ * Chair of Geoinformatics
+ * Technical University of Munich, Germany
+ * https://www.lrg.tum.de/gis/
+ *
+ * The 3D City Database is jointly developed with the following
+ * cooperation partners:
+ *
+ * Virtual City Systems, Berlin <https://vc.systems/>
+ * M.O.S.S. Computer Grafik Systeme GmbH, Taufkirchen <http://www.moss.de/>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.citydb.core.file.input;
+
+import org.apache.tika.exception.TikaException;
+import org.apache.tika.mime.MediaType;
+import org.citydb.core.file.InputFile;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URI;
+import java.nio.file.*;
+import java.util.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Stream;
+import java.util.zip.GZIPInputStream;
+
+
+public class FileProcessor {
+    private final FileTypeDetector fileTypeDetector;
+    private final Pattern contentFilePattern;
+    private final Matcher matcher;
+
+
+    public FileProcessor(Pattern contentFilePattern) throws TikaException, IOException {
+        this.fileTypeDetector = new FileTypeDetector();
+        this.contentFilePattern = contentFilePattern;
+        this.matcher = Pattern.compile("").matcher("");
+    }
+
+    public void processFile(Path file, List<InputFile> files, boolean force) throws IOException {
+        MediaType mediaType = fileTypeDetector.getMediaType(file);
+        if (mediaType.equals(InputFile.APPLICATION_ZIP)) {
+            processZipFile(file, files);
+        } else if (mediaType.equals(InputFile.APPLICATION_GZIP)) {
+            processGZipFile(file, files);
+        } else if (isSupportedContentType(mediaType) || force) {
+            files.add(new RegularInputFile(file, mediaType));
+        }
+    }
+
+    private void processGZipFile(Path gzipFile, List<InputFile> files) {
+        try (InputStream stream = new GZIPInputStream(new FileInputStream(gzipFile.toFile()))) {
+            // pass file name without gzip extension as hint for content detection
+            String fileName = gzipFile.getFileName().toString();
+            fileName = deriveFileNameWithoutGzipExtension(fileName);
+            MediaType mediaType = fileTypeDetector.getMediaType(stream, fileName);
+            if (isSupportedContentType(mediaType)) {
+                files.add(new GZipInputFile(gzipFile, mediaType));
+            }
+        } catch (IOException e) {
+            //
+        }
+    }
+
+    private void processZipFile(Path zipFile, List<InputFile> files) throws IOException {
+        URI uri = URI.create("jar:" + zipFile.toAbsolutePath().toUri());
+        FileSystem fileSystem = null;
+        try {
+            fileSystem = FileSystems.getFileSystem(uri);
+        } catch (FileSystemNotFoundException e) {
+            //
+        } catch (Throwable e) {
+            throw new IOException("Failed to open zip file system '" + uri + "'.", e);
+        }
+        if (fileSystem == null) {
+            try {
+                fileSystem = FileSystems.newFileSystem(uri, new HashMap<>());
+            } catch (Throwable e) {
+                throw new IOException("Failed to open zip file system '" + uri + "'.", e);
+            }
+        }
+        try (Stream<Path> stream = Files.walk(fileSystem.getPath("/")).filter(Files::isRegularFile)) {
+            stream.forEach(path -> {
+                matcher.reset(path.getFileName().toString()).usePattern(this.contentFilePattern);
+                if (matcher.matches()) {
+                    MediaType mediaType = fileTypeDetector.getMediaType(path);
+                    if (isSupportedContentType(mediaType)) {
+                        ZipInputFile zipInputFile = new ZipInputFile(path.toString(), zipFile, uri, mediaType);
+                        files.add(zipInputFile);
+                    }
+                }
+            });
+        } catch (Throwable e) {
+            throw new IOException("Failed to read zip file entries.", e);
+        } finally {
+            fileSystem.close();
+        }
+    }
+
+    private boolean isSupportedContentType(MediaType mediaType) {
+        return mediaType.equals(InputFile.APPLICATION_XML)
+                || mediaType.equals(InputFile.APPLICATION_JSON);
+    }
+
+    private String deriveFileNameWithoutGzipExtension(String fileName) {
+        if (fileName.endsWith(".gz")) {
+            return fileName.substring(0, fileName.length() - 3);
+        } else if (fileName.endsWith(".gzip")) {
+            return fileName.substring(0, fileName.length() - 5);
+        }
+        return fileName;
+    }
+}
